Fix #10538: Switch cassandra-driver to a proper module, and add a tsconfig.json
